Fix merging
